History screen:
- Home3 string improvements
- Humanizer.plural()

+++ library/src/main/java/es/rafaco/inappdevtools/library/storage/db/entities/NetSummaryDao.java
@@ -69,7 +69,7 @@ public interface NetSummaryDao {
     long sizeBySession(long sessionId);
 
     @Query("SELECT COUNT(*) FROM net_summary WHERE sessionId LIKE :sessionId and status LIKE :status")
-    long countBySessionAndStatus(long sessionId, int status);
+    int countBySessionAndStatus(long sessionId, int status);
 
     @Insert
     long insert(NetSummary netsummary);
